make the array stuff work, in theory.
i think this should work now.  GArrays of GValues don't seem to work,
unfortunately, but GPtrArrays of GValues do.  passing GValueArrays
directly to dbus-glib works, but it's difficult if not impossible to
figure out what they are on the other end.

anyhow, so the array stuff works by passing a GPtrArray of GValues
to dbus-glib (after stuffing the GPtrArray in its own GValue).


(Old svn revision: 26774)