add property value to PropertyChanged/XfconfChannel::property-changed signal
it seems like every time i get a property-changed signal, the first thing
i do is go and fetch the property.  always sending the value over the wire
in the signal will generally save us a round-trip when handling property
changes.  the downside is that the value always gets sent out on any prop
change, regardless if anyone cares about that particular property or not


(Old svn revision: 26876)