add transparent property prefetching and caching
XfconfCache is a transparent cache that sits behind XfconfChannel,
and, so far, involves no new API.  this moves most of the dbus code
from XfconfChannel to XfconfCache.  on creation, XfconfChannel gets
an XfconfCache object for that channel (the cache objects are
per-channel-name singletons, regardless of property_base or whether
or not the XfconfChannel is a singleton).

on creation, XfconfChannel prefetches all properties in the channel.
this might turn out to be a bad idea for large channels.

property setting does not block if the setting is already in the cache.
i may change this in the future to not block at all ever, but that
might make reporting certain kinds of errors impossible.  property
resets are completely non-blocking.  lookups are non-blocking if
the setting is already in the cache.  the cache automatically
updates itself if another application modifies a property.

the cache has hooks to set the maximum age of entries (in seconds)
and the max number of entries to store in the cache.  currently
these two are unimplemented, and i'm not sure if there's value to
exposing these in the public API.

(Old svn revision: 30417)