New internal (simpler) handling of bindings.
Code is only using signals for connecting and disconnecting the
bindings. This works better with disconnecting from xfconf
using weak refs.

It also uses one internal list for storing all the connected
bindings, instead of a hash table and a list connected to the
object. This makes it easier to understand and disconnecting
is not really a performance issues here since we don't do
that very often (I still think this is faster since the
previous implementation poked both the hash table and the
list attached to the object).

For the binding id, it uses the the signal handler from the
channel's property-changed signal, which allows to drop the
internal id counter.

Forthermore some warnings are printed when bindings are still
connected when xfconf_shutdown() is called or when
an unbind function is called without any result.