Fix segfault triggered by object-bindings test.
It turned out the object bindings code triggered a bug
in the cache code because it edits multiple properties
with the same name in a short time and then destroying
the channel.
This left a number (6) of pending call keys in one of the
hash tables with the same value resulting in double frees
when destroying the table.

To fix this steal the item from the table when cancelling
the dbus call, then the old_item we insert again is then
not freed and exists only once in the table.