Wait for pending calls in the cache.
Brian already added a FIXME with the question if we should
wait for pending calls to finish before freeing the items.
Well the question is yes, because the get-tests failed on my
system because the channel was closed in the set-tests before
dbus replied and thus nothing was stored in the test-channel.

So wait for pending calls in finalize and skip the normal reply
handler by setting the hash table in the cache to NULL.