Use glibs implementation for loading mapped files.
Valgrind reported a lot of memory corruption in this part of
the code, so rewrote it a bit to use the glib functions that
are provided for this task (with fallback to normal file loading
if glib has no mmap support).

The old mmap code was also never enabled because of missing includes
and checks in configure.ac.in.