Don't use cache singletons.
None of the Xfce apps opens more then 2 new channels with
the same name, so that means we never return an existing
cache object and use it twice. It also leaked the hash table
for the cache singletons, so optimize for the 'common case'.

2 of the same caches works fine (but doubles the dbus traffic)
so to really fix this an app should use the channel singletons.
Mentioned this in the API docs.