Fix double free crash (bug #8169).
In xfconf_cache_set_property_reply_handler() if the item is not found in
cache->properties, the function exit (goto out;) without removing the
old_property from cache->old_properties nor the call from cache->pending_calls.

Then when xfconf_cache_set() is called, the old_item is still found in the hash
(as it wasn't removed previously) and therefore dbus_g_proxy_cancel_call() is
called in a call which was completed, thus leading to the double-free and the
crash.