Allow :.,[]{} characters in property and channel names (bug #11021)
Among other things, this fixes an xfdesktop issue with NVIDIA cards,
where connected monitors are reported as "monitorGPU-0.DVI-I-3" instead
of "monitor0". By allowing dots and other characters in property and
channel names, these card/monitor combinations are now supported
properly.

This may also be useful if we make xfconf a gsettings backend in the
future.