Use a custom test-driver
Implement a custom test driver that will start the freshly compiled
xfconfd on a different dbus name. When the test finishes, gets
interrupted or exists with error, the xfconfd test instance is
terminated.

This will allows 'make check' and 'make distcheck' to be successful even
if xfconfd is not installed or running.