ci: Upload artifacts on failure without specifying a directory
`on_failure` allows to retrieve artifacts in case of failure, and only
in that case: see
https://docs.gitlab.com/ee/ci/yaml/index.html#artifactswhen

The build takes place in a directory of the form
`xfconf-${version}git${commit}/_build/sub` at the moment, but the
`_build/sub` part could possibly change in the future, as it is added by
Gitlab. So the easiest and most robust way seems to be to get all files
named `test-suite.log`.