Allow calling xfconf_g_property_bind() with GdkColor or GdkRGBA types
There's really no reason for the "special case" public APIs for binding
GdkColor and GdkRGBA types: we can simply check for these in
xfconf_g_property_bind() and handle them specially.

This is helpful for cases where you have a bunch of properties to bind,
and have listed them in a struct that matches setting names and types
with gobject property names.  That way you don't have to special-case
any GdkRGBA or GdkColor types when creating the binding.