channel: Fix get_arrayv() return value
When the array is empty, it should be returned as is, instead of NULL:
this is not a case of failure. So, in particular, when get_string_list()
is called, an empty string array (NULL-terminated) is returned, and not
NULL, as advertised in its documentation: get_string_list() should only
return NULL if the property doesn't exist (or in case of failure more
generally).